+++ resources/js/api/base.js
@@ -48,6 +48,8 @@ export default {
       vm.$router.replace({ name: 'home' });
     } else if (responseStatus === env.HTTP_PAYLOAD_TOO_LARGE) { // 413 => payload to large
       vm.flashMessage.error(vm.$t('app.flash.tooLarge'));
+    } else if (responseStatus === env.HTTP_SERVICE_UNAVAILABLE) { // 503 => maintenance mode
+      window.location.reload();
     } else if (responseStatus !== undefined) { // Another error on server
       vm.flashMessage.error({
         message: errorMessage ? vm.$t('app.flash.serverError.message', { message: errorMessage }) : vm.$t('app.flash.serverError.emptyMessage'),