+++ tests/Feature/api/v1/Room/RoomTest.php
@@ -347,7 +347,8 @@ class RoomTest extends TestCase
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index'))
             ->assertStatus(200)
             ->assertJsonCount(1, 'data')
-            ->assertJsonFragment(['id'=>$room5->id,'name'=>$room5->name]);
+            ->assertJsonFragment(['id'=>$room5->id,'name'=>$room5->name])
+            ->assertJsonCount(7, 'meta');
 
         // Test with viewAll rooms permission
         $role       = factory(Role::class)->create();
@@ -359,12 +360,15 @@ class RoomTest extends TestCase
             ->assertJsonFragment(['id'=>$room1->id,'name'=>$room1->name])
             ->assertJsonFragment(['id'=>$room2->id,'name'=>$room2->name])
             ->assertJsonFragment(['id'=>$room3->id,'name'=>$room3->name])
-            ->assertJsonFragment(['id'=>$room4->id,'name'=>$room4->name]);
+            ->assertJsonFragment(['id'=>$room4->id,'name'=>$room4->name])
+            ->assertJsonCount(7, 'meta');
 
         // Find by room name
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?search=test')
             ->assertStatus(200)
-            ->assertJsonCount(2, 'data');
+            ->assertJsonCount(2, 'data')
+            ->assertJsonCount(7, 'meta');
+
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?search=test+a')
             ->assertStatus(200)
             ->assertJsonCount(1, 'data');
@@ -394,7 +398,9 @@ class RoomTest extends TestCase
         // Filter by room types
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?roomTypes[]='.$roomType1->id)
             ->assertStatus(200)
-            ->assertJsonCount(2, 'data');
+            ->assertJsonCount(2, 'data')
+            ->assertJsonCount(7, 'meta');
+
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?roomTypes[]='.$roomType1->id.'&roomTypes[]='.$roomType2->id)
             ->assertStatus(200)
             ->assertJsonCount(4, 'data');
@@ -435,12 +441,17 @@ class RoomTest extends TestCase
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?filter=own')
             ->assertOk()
             ->assertJsonFragment(['id'=>$rooms[2]->id])
-            ->assertJsonFragment(['id'=>$rooms[3]->id]);
+            ->assertJsonFragment(['id'=>$rooms[3]->id])
+            ->assertJsonPath('meta.total', 2)
+            ->assertJsonPath('meta.total_no_filter', 2)
+            ->assertJsonCount(8, 'meta');
 
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?filter=shared')
             ->assertOk()
             ->assertJsonFragment(['id'=>$rooms[0]->id])
-            ->assertJsonFragment(['id'=>$rooms[1]->id]);
+            ->assertJsonFragment(['id'=>$rooms[1]->id])
+            ->assertJsonPath('meta.total', 2)
+            ->assertJsonPath('meta.total_no_filter', 2);
     }
 
     /**
@@ -457,27 +468,37 @@ class RoomTest extends TestCase
         // Testing without query
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?filter=own')
             ->assertOk()
-            ->assertJsonFragment(['id'=>$room->id]);
+            ->assertJsonFragment(['id'=>$room->id])
+            ->assertJsonPath('meta.total', 1)
+            ->assertJsonPath('meta.total_no_filter', 1);
 
         // Testing with empty query
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?filter=own&search=')
             ->assertOk()
-            ->assertJsonFragment(['id'=>$room->id]);
+            ->assertJsonFragment(['id'=>$room->id])
+            ->assertJsonPath('meta.total', 1)
+            ->assertJsonPath('meta.total_no_filter', 1);
 
         // Testing with fragment of the room name
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?filter=own&search=One')
             ->assertOk()
-            ->assertJsonFragment(['id'=>$room->id]);
+            ->assertJsonFragment(['id'=>$room->id])
+            ->assertJsonPath('meta.total', 1)
+            ->assertJsonPath('meta.total_no_filter', 1);
 
         // Testing with full name
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?filter=own&search=Meeting One')
             ->assertOk()
-            ->assertJsonFragment(['id'=>$room->id]);
+            ->assertJsonFragment(['id'=>$room->id])
+            ->assertJsonPath('meta.total', 1)
+            ->assertJsonPath('meta.total_no_filter', 1);
 
         // Testing with invalid name
         $this->actingAs($this->user)->getJson(route('api.v1.rooms.index').'?filter=own&search=Meeting Two')
             ->assertOk()
-            ->assertJsonMissing(['id'=>$room->id]);
+            ->assertJsonMissing(['id'=>$room->id])
+            ->assertJsonPath('meta.total', 0)
+            ->assertJsonPath('meta.total_no_filter', 1);
     }
 
     /**