+++ tests/Frontend/views/rooms/View.spec.js
@@ -11,6 +11,7 @@ import VueRouter from 'vue-router';
 import PermissionService from '../../../../resources/js/services/PermissionService';
 import Vue from 'vue';
 import _ from 'lodash';
+import env from '../../../../resources/js/env';
 
 const localVue = createLocalVue();
 
@@ -449,5 +450,30 @@ describe('Room', function () {
     Base.error.restore();
     handleInvalidCode.restore();
     view.destroy();
+  });
+
+  it('error beforeRouteEnter', function (done) {
+    moxios.stubRequest('/api/v1/rooms/abc-def-456', {
+      status: env.HTTP_SERVICE_UNAVAILABLE
+    });
+
+    const view = mount(RoomView, {
+      localVue,
+      mocks: {
+        $t: (key) => key
+      },
+      store,
+      attachTo: createContainer()
+    });
+
+    const to = { params: { id: 'abc-def-456' } };
+
+    const next = (error) => {
+      expect(error instanceof Error).toBeTruthy();
+      expect(error.response.status).toBe(env.HTTP_SERVICE_UNAVAILABLE);
+      view.destroy();
+      done();
+    };
+    RoomView.beforeRouteEnter.call(view.vm, to, undefined, next);
   });
 });