+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Incorrect count of user rooms if search is used ([#129], [#130])
+- Broken error handling on room view navigation ([#131], [#132])
 
 ### Added
 - Co-owner room role, permissions to view and edit all rooms (incl. memberships, files and settings) ([#110], [#116])
@@ -125,6 +128,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#122]: https://github.com/THM-Health/PILOS/pull/122
 [#123]: https://github.com/THM-Health/PILOS/issues/123
 [#125]: https://github.com/THM-Health/PILOS/pull/125
+[#129]: https://github.com/THM-Health/PILOS/issues/129
+[#130]: https://github.com/THM-Health/PILOS/pull/130
+[#131]: https://github.com/THM-Health/PILOS/issues/131
+[#132]: https://github.com/THM-Health/PILOS/pull/132
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v1.1.0...HEAD
 [1.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v1.0.0